url-content-extraction/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
selenium>=4.0.0
helium>=3.0.0
transformers>=4.0.0
huggingface-hub>=0.0.1
beautifulsoup4-4.13.3
bs4-0.0.2 
soupsieve-2.6
pillow-11.1.0
